adding updates for front end login, sign up, profile and welcome
