+++ apps/web/src/screens/Login.js
@@ -1,14 +1,36 @@
+
 import React, { useState } from "react";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import logo from "../assets/cardconnect-logo.png";
 
 export default function Login() {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [error, setError] = useState("");
+  const navigate = useNavigate();
 
-  const handleSubmit = (e) => {
+  const handleSubmit = async (e) => {
     e.preventDefault();
-    alert(`Logging in with Email: ${email}`);
+    setError("");
+
+    try {
+      const response = await fetch("http://localhost:8080/api/admin/login", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ email, password }),
+      });
+
+      if (!response.ok) {
+        const data = await response.json();
+        throw new Error(data.message || "Login failed");
+      }
+
+      const data = await response.json();
+      localStorage.setItem("user", JSON.stringify(data));
+      navigate("/profile");
+    } catch (err) {
+      setError(err.message);
+    }
   };
 
   const containerStyle = {
@@ -75,6 +97,12 @@ export default function Login() {
     textAlign: "center",
   };
 
+  const errorStyle = {
+    color: "red",
+    fontSize: "14px",
+    marginTop: "10px",
+  };
+
   return (
     <div style={containerStyle}>
       {/* Left Side */}
@@ -82,11 +110,7 @@ export default function Login() {
         <img
           src={logo}
           alt="Card Connect Logo"
-          style={{
-            width: "80px",
-            height: "80px",
-            marginBottom: "15px",
-          }}
+          style={{ width: "80px", height: "80px", marginBottom: "15px" }}
         />
         <h1 style={{ fontSize: "28px", marginBottom: "10px" }}>
           Card Connect Login
@@ -122,8 +146,10 @@ export default function Login() {
             Login
           </button>
 
+          {error && <div style={errorStyle}>{error}</div>}
+
           <div style={smallTextStyle}>
-            Don't have an account?{" "}
+            Don&apos;t have an account?{" "}
             <Link to="/signup" style={linkStyle}>
               Sign up here
             </Link>
